+++ Sources/Glossa/Services/SystemApplicationService.swift
@@ -11,15 +11,15 @@ protocol SystemApplicationServing: AnyObject {
 final class SystemApplicationService: SystemApplicationServing {
     func openSystemAudioPermissionSettings() {
         openSystemSettingsPane([
-            "x-apple.systempreferences:com.apple.preference.security?Privacy_ScreenCapture",
-            "x-apple.systempreferences:com.apple.settings.PrivacySecurity.extension?Privacy_ScreenCapture"
+            "x-apple.systempreferences:com.apple.settings.PrivacySecurity.extension?Privacy_ScreenCapture",
+            "x-apple.systempreferences:com.apple.preference.security?Privacy_ScreenCapture"
         ])
     }
 
     func openMicrophonePermissionSettings() {
         openSystemSettingsPane([
-            "x-apple.systempreferences:com.apple.preference.security?Privacy_Microphone",
-            "x-apple.systempreferences:com.apple.settings.PrivacySecurity.extension?Privacy_Microphone"
+            "x-apple.systempreferences:com.apple.settings.PrivacySecurity.extension?Privacy_Microphone",
+            "x-apple.systempreferences:com.apple.preference.security?Privacy_Microphone"
         ])
     }